little

+++ includes/ft_ls.h
@@ -6,7 +6,7 @@
 /*   By: timofieiev <timofieiev@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/20 15:18:36 by otimofie          #+#    #+#             */
-/*   Updated: 2018/11/11 00:45:26 by timofieiev       ###   ########.fr       */
+/*   Updated: 2018/11/11 00:59:43 by timofieiev       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ typedef	struct		s_temp
 	unsigned char	d_type;
 	long long int		blocks_data;
 	char			type_and_permissions_data[12];
-	unsigned int	hard_links_data;
+	long int	hard_links_data;
 	char			owner_name_data[20];
 	char			group_name_data[20];
 	long long int	size_data;
@@ -54,7 +54,7 @@ typedef	struct		s_data
 {
 	long long	int	blocks_buf;
 	char			type_and_permissions_buf[12];
-	unsigned int	hard_links_buf;
+	long  int	hard_links_buf;
 	char			owner_name_buf[20]; 
 	char			group_name_buf[20];
 	long long int	size_buf;
