+++ srcs/stats_infrastructure.c
@@ -6,7 +6,7 @@
 /*   By: timofieiev <timofieiev@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/28 13:38:08 by otimofie          #+#    #+#             */
-/*   Updated: 2018/11/11 00:41:13 by timofieiev       ###   ########.fr       */
+/*   Updated: 2018/11/11 01:00:36 by timofieiev       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,8 +63,9 @@ t_data	get_stats(char *buffer_inner)
 	get_permissions(stats.type_and_permissions_buf, buf.st_mode);
 	get_xattr(stats.type_and_permissions_buf, buffer_inner);
 	stats.type_and_permissions_buf[11] = '\0';
-	stats.hard_links_buf = buf.st_nlink;
+	stats.hard_links_buf = (long int)buf.st_nlink;
 	stats.size_buf = (long long int)buf.st_size;
+	// ft_printf("---%lld---",(long long int)buf.st_size);
 	buf1 = getpwuid(buf.st_uid)->pw_name;
 	str_copy(stats.owner_name_buf, buf1);
 	buf1 = getgrgid(buf.st_gid)->gr_name;
